fix services

+++ scripts/install-services.ps1
@@ -1,7 +1,11 @@
 # Disable any services that we installed
 $OpenVPNServiceName = 'OpenVPNServiceInteractive'
-Stop-Service -Name $OpenVPNServiceName
-Set-Service -Name $OpenVPNServiceName -StartupType Disabled
+$service = Get-Service -Name W32Time -ErrorAction SilentlyContinue
+if ($service.Length -gt 0) {
+    Stop-Service -Name $OpenVPNServiceName
+    Set-Service -Name $OpenVPNServiceName -StartupType Disabled
+}
+
 # Delete KeePassXC startup key
 $KeepassXC_Path = 'HKCU:\Software\Microsoft\Windows\CurrentVersion\Run'
 $KeepassXC_Name = 'KeePassXC'
